Add files via upload
The motive of the project is to create a basic Report in Power BI and showcasing my Skills of using the software correctly. I tried Finding some answers from the data sheet which would be helpful in finding the insights of. Questions such as:
Finding out the month on month basis Revenue and Profits made In 2018, Creating a slicer for Quarter wise analysis.
Finding out which subcategory is giving us the most profit and which is giving us the most loss. 
Looking at which is the most preferred payment mode and analysing
Finding Top 5 states with the most profit.
Finding out the sum of categories(top 3 only)  and sub-categories on the basis of Quantity.

By looking at these insights we can make major decisions like what is our most profitable side, which side we need to focus more/what could be the reason for having poor performance in a particular Category/State.